+++ .gitignore
@@ -9,3 +9,8 @@ dist/
 *.db
 venv/
 uv.lock
+
+# local SDK test-run artifacts (generated by demos; keep tracked example json)
+examples/workspace/files/
+examples/workspace/*.db
+examples/documents/attention.pdf
