--- pageindex/index/legacy_utils.py
@@ -1,2 +1,0 @@
-# Re-export from the original utils.py for backward compatibility
-from ..utils import *
