+++ pageindex/retrieve.py
@@ -2,9 +2,9 @@ import json
 import PyPDF2
 
 try:
-    from .utils import get_number_of_pages, remove_fields
+    from .index.utils import get_number_of_pages, remove_fields
 except ImportError:
-    from utils import get_number_of_pages, remove_fields
+    from index.utils import get_number_of_pages, remove_fields
 
 
 # ── Helpers ──────────────────────────────────────────────────────────────────
